moved draw grid method  to generic class
